cost-reporter/requirements.txt
numpy
matplotlib
slackclient
discord.py==2.3.2
boto3